requirements.txt
Flask==2.0.1
anthropic==0.3.0
beautifulsoup4==4.9.3
requests==2.26.0
python-dotenv==0.19.0
playwright==1.14.0
python-dotenv